+++ test/multiapi/Expected/AcceptanceTests/MultiapiWithSubmodule/multiapiwithsubmodule/submodule/v3/aio/operations_async/_operation_group_two_operations_async.py
@@ -74,9 +74,9 @@ class OperationGroupTwoOperations:
 
         # Construct and send request
         body_content_kwargs = {}  # type: Dict[str, Any]
-        if header_parameters['Content-Type'] in ['application/pdf', 'image/jpeg', 'image/png', 'image/tiff']:
+        if header_parameters['Content-Type'].split(";")[0] in ['application/pdf', 'image/jpeg', 'image/png', 'image/tiff']:
             body_content_kwargs['stream_content'] = input
-        elif header_parameters['Content-Type'] in ['application/json']:
+        elif header_parameters['Content-Type'].split(";")[0] in ['application/json']:
             if input is not None:
                 body_content = self._serialize.body(input, 'SourcePath')
             else:
@@ -84,7 +84,8 @@ class OperationGroupTwoOperations:
             body_content_kwargs['content'] = body_content
         else:
             raise ValueError(
-                "Content type {} is not valid for this operation".format(header_parameters['Content-Type'])
+                "The content_type '{}' is not one of the allowed values: "
+                "['application/pdf', 'image/jpeg', 'image/png', 'image/tiff', 'application/json']".format(header_parameters['Content-Type'])
             )
         request = self._client.post(url, query_parameters, header_parameters, **body_content_kwargs)