+++ ChangeLog.md
@@ -1,12 +1,13 @@
 # Change Log
 
-### Unreleased
-Modelerfour version: 4.13.346
+### 2020-05-08 - 5.0.0-preview.5
+Modelerfour version: 4.13.351
 
 **Bug Fixes**
 
 - Users can pass in content types with ';' inside (such as 'text/plain; encoding=UTF-8')  #619
 - Allowing parameters to be of type `IO` as well  #618
+- Can now generate without FATAL: bad indentation error (taken from m4 update - perks PR #105)
 
 ### 2020-05-06 - 5.0.0-preview.4
 Modelerfour version: 4.13.346